backend/requirements.txt
beautifulsoup4>=4.14.2
docling>=2.62.0
fastapi>=0.121.3
feedparser>=6.0.12
google-genai>=1.51.0
markdown>=3.10
markdownify>=1.2.2
psycopg2-binary>=2.9.11
pydantic[email]>=2.12.4
python-dotenv>=1.2.1
python-multipart>=0.0.20
requests>=2.32.5
sqlalchemy>=2.0.44
youtube-transcript-api>=1.2.3
